Pagination (#44)
* Add timeout to failed login message

* Update generated API

* Create Paginator component

* Add Paginator to LeaderboardsList page

* Add pagination when fetching categories

* Fix old API calls

* Rename query param

* Formatting

* Actually make limit selector work

* Move position of Paginator to prevent UI bug
Calling Paginator outside of the v-else block causes it to display 0 in
the select, and it won't update with a change of the prop value. It
needs to be called in the right position in the parent component to
prevent this bug from happening.

* Add tabindex to improve a11y

* Remove anchor tab focus in button links
Prevents button links from being focused on twice